implimented modals and fixed bad routing paths to resolve single page rendering issues
