+++ src/components/edit-todo.comp.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import axios from "axios";
-import {Form, Col, Row, Button} from "react-bootstrap";
-
+import { Link } from "react-router-dom";
+import { Form, Col, Row, Button, Modal } from "react-bootstrap";
 
 export default class EditTodo extends Component {
   constructor(props) {
@@ -13,11 +13,15 @@ export default class EditTodo extends Component {
     this.onChangeTodoCompleted = this.onChangeTodoCompleted.bind(this);
     this.onSubmit = this.onSubmit.bind(this);
 
+    this.close = this.close.bind(this);
+    this.open = this.open.bind(this);
+
     this.state = {
       todo_description: "",
       todo_responsible: "",
       todo_priority: "",
-      todo_completed: false
+      todo_completed: false,
+      showModal: true
     };
   }
 
@@ -36,6 +40,17 @@ export default class EditTodo extends Component {
         console.log(error);
       });
   }
+
+  // componentDidUpdate() {
+  //   axios
+  //     .get("http://localhost:4000/todos/")
+  //     .then(response => {
+  //       this.setState({ todos: response.data });
+  //     })
+  //     .catch(function(error) {
+  //       console.log(error);
+  //     });
+  // }
 
   onChangeTodoDescription(e) {
     this.setState({
@@ -63,6 +78,7 @@ export default class EditTodo extends Component {
 
   onSubmit(e) {
     e.preventDefault();
+
     const obj = {
       todo_description: this.state.todo_description,
       todo_responsible: this.state.todo_responsible,
@@ -80,89 +96,108 @@ export default class EditTodo extends Component {
     this.props.history.push("/");
   }
 
+  close() {
+    this.setState({ showModal: false });
+  }
+
+  open() {
+    this.setState({ showModal: true });
+  }
+
   render() {
     return (
       <div>
-        <h3 align="center">Update Todo</h3>
-        <Form onSubmit={this.onSubmit}>
-          <Form.Group className="form-group">
-            <Form.Label>Description: </Form.Label>
-            <Form.Control
-              type="text"
-              placeholder="What are we doing today?"
-              className="form-control"
-              value={this.state.todo_description}
-              onChange={this.onChangeTodoDescription}
-            />
-            <Form.Text className="text-muted">
-              {/* We'll never share your email with anyone else. */}
-            </Form.Text>
-          </Form.Group>
+        <Modal show={this.state.showModal} onHide={this.close}>
+          <Modal.Header>
+            <Modal.Title>Edit Task</Modal.Title>
+          </Modal.Header>
+          <Modal.Body>
+            <Form onSubmit={this.onSubmit}>
+              <Form.Group className="form-group">
+                <Form.Label>Description: </Form.Label>
+                <Form.Control
+                  type="text"
+                  placeholder="What are we doing today?"
+                  className="form-control"
+                  value={this.state.todo_description}
+                  onChange={this.onChangeTodoDescription}
+                />
+                <Form.Text className="text-muted">
+                  {/* We'll never share your email with anyone else. */}
+                </Form.Text>
+              </Form.Group>
 
-          <Form.Group className="form-group">
-            <Form.Label>Responsible: </Form.Label>
-            <Form.Control
-              type="text"
-              placeholder="Who will be doing this today?"
-              className="form-control"
-              value={this.state.todo_responsible}
-              onChange={this.onChangeTodoResponsible}
-            />
-          </Form.Group>
-          {/* fieldset is like a div but used as a way to group */}
-          <fieldset>
-            <Form.Group className="form-group" as={Row}>
-              <Form.Label as="legend" column sm={2}>
-                Priority
-              </Form.Label>
-              <Col sm={10}>
-                <Form.Check
-                  type="radio"
-                  label="Low"
-                  name="priorityOptions"
-                  id="priorityLow"
-                  value="Low"
-                  checked={this.state.todo_priority === "Low"}
-                  onChange={this.onChangeTodoPriority}
+              <Form.Group className="form-group">
+                <Form.Label>Responsible: </Form.Label>
+                <Form.Control
+                  type="text"
+                  placeholder="Who will be doing this today?"
+                  className="form-control"
+                  value={this.state.todo_responsible}
+                  onChange={this.onChangeTodoResponsible}
                 />
-                <Form.Check
-                  type="radio"
-                  label="Medium"
-                  name="priorityOptions"
-                  id="priorityMedium"
-                  value="Medium"
-                  checked={this.state.todo_priority === "Medium"}
-                  onChange={this.onChangeTodoPriority}
-                />
-                <Form.Check
-                  type="radio"
-                  label="High"
-                  name="priorityOptions"
-                  id="priorityHigh"
-                  value="High"
-                  checked={this.state.todo_priority === "High"}
-                  onChange={this.onChangeTodoPriority}
-                />
-              </Col>
-            </Form.Group>
-            <Form.Group>
-              <Form.Check
-                id="completedCheckbox"
-                type="checkbox"
-                name="completedCheckbox"
-                onChange={this.onChangeTodoCompleted}
-                checked={this.state.todo_completed}
-                value={this.state.todo_completed}
-                label="Completed"
-                htmlFor="completedCheckbox"
-                />
-            </Form.Group>
-          </fieldset>
+              </Form.Group>
+              {/* fieldset is like a div but used as a way to group */}
+              <fieldset>
+                <Form.Group className="form-group" as={Row}>
+                  <Form.Label as="legend" column sm={2}>
+                    Priority
+                  </Form.Label>
+                  <Col sm={10}>
+                    <Form.Check
+                      type="radio"
+                      label="Low"
+                      name="priorityOptions"
+                      id="priorityLow"
+                      value="Low"
+                      checked={this.state.todo_priority === "Low"}
+                      onChange={this.onChangeTodoPriority}
+                    />
+                    <Form.Check
+                      type="radio"
+                      label="Medium"
+                      name="priorityOptions"
+                      id="priorityMedium"
+                      value="Medium"
+                      checked={this.state.todo_priority === "Medium"}
+                      onChange={this.onChangeTodoPriority}
+                    />
+                    <Form.Check
+                      type="radio"
+                      label="High"
+                      name="priorityOptions"
+                      id="priorityHigh"
+                      value="High"
+                      checked={this.state.todo_priority === "High"}
+                      onChange={this.onChangeTodoPriority}
+                    />
+                  </Col>
+                </Form.Group>
+                <Form.Group>
+                  <Form.Check
+                    id="completedCheckbox"
+                    type="checkbox"
+                    name="completedCheckbox"
+                    onChange={this.onChangeTodoCompleted}
+                    checked={this.state.todo_completed}
+                    value={this.state.todo_completed}
+                    label="Completed"
+                    htmlFor="completedCheckbox"
+                  />
+                </Form.Group>
+              </fieldset>
 
-          <Button variant="warning" type="submit" value="Update Todo">
-            Update Todo
-          </Button>
-        </Form>
+              <Button variant="warning" type="submit" value="Update Todo">
+                Update Task
+              </Button>
+            </Form>
+          </Modal.Body>
+          <Modal.Footer>
+            <Link to="/todos">
+              <Button variant="secondary">Close</Button>
+            </Link>
+          </Modal.Footer>
+        </Modal>
       </div>
     );
   }
