+++ src/components/modal-create-todo.comp.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { Modal, Form, Col, Row, Button, Table } from "react-bootstrap";
+import { Modal, Button} from "react-bootstrap";
 import CreateToDo from "./create-todo.comp";
